// src/apps/interfaces/AppInterface.ts
export interface FolderInterface {
    name: string;
    icon?: string;
    defaultFilesIcon?: string;
    condition?: () => boolean | Promise<boolean>;
    action?: () => void;
    pathName?: string;
    pathUrl?: string;
    files?: FolderInterface[]
}